fix game endpoinds & update gameround table

+++ src/main/java/com/musiguessr/backend/dto/post/RoundSummaryDTO.java
@@ -11,7 +11,7 @@ public class RoundSummaryDTO {
     private boolean response;
     private String songName;
     private String guessedSong;
-    private Integer guessTime;
+    private Long guessTime;
     private Integer scoreEarned;
     private Integer round;
 }
